# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Production-style REST API server built with Go.",
        "title": "Chirpy API",
        "termsOfService": "https://github.com/yourusername/chirpy",
        "contact": {
            "name": "Mirodilbek Fazilov",
            "url": "https://github.com/odilmode"
        },
        "license": {
            "name": "MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/admin/reset": {
            "post": {
                "description": "Deletes all users from the database and resets hit counter. Only accessible in development environment.",
                "produces": [
                    "text/plain"
                ],
                "tags": [
                    "admin"
                ],
                "summary": "Reset users and file server hits",
                "responses": {
                    "200": {
                        "description": "All users deleted and hits reset to 0",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "403": {
                        "description": "Forbidden: This endpoint is only accessible in development environment",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Failed to delete all users",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/chirps": {
            "get": {
                "description": "Retrieve chirps, optionally filtered by author_id and sorted by creation time",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "chirps"
                ],
                "summary": "Get Chirps",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Filter chirps by author UUID",
                        "name": "author_id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Sort order: asc (default) or desc",
                        "name": "sort",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/main.Chirp"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid author_id",
                        "schema": {
                            "$ref": "#/definitions/main.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Failed to fetch chirps or encode response",
                        "schema": {
                            "$ref": "#/definitions/main.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Authenticated endpoint to create a chirp with max length 140 characters. Filters bad words.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "chirps"
                ],
                "summary": "Create a new chirp",
                "parameters": [
                    {
                        "description": "Chirp body",
                        "name": "chirp",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/main.requestBody"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/main.Chirp"
                        }
                    },
                    "400": {
                        "description": "Invalid author_id",
                        "schema": {
                            "$ref": "#/definitions/main.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized - missing or invalid JWT",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error - failed to create chirp",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
        "/api/polka/webhooks": {
            "post": {
                "description": "Handles webhook events from Polka, such as user upgrade notifications",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "webhooks"
                ],
                "summary": "Handle Polka Webhooks",
                "parameters": [
                    {
                        "type": "string",
                        "description": "API Key for authentication",
                        "name": "X-API-Key",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "Webhook event payload",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "properties": {
                                "data": {
                                    "type": "object",
                                    "properties": {
                                        "user_id": {
                                            "type": "string"
                                        }
                                    }
                                },
                                "event": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Invalid request or user ID",
                        "schema": {
                            "$ref": "#/definitions/main.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized: invalid or missing API key",
                        "schema": {
                            "$ref": "#/definitions/main.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "User not found",
                        "schema": {
                            "$ref": "#/definitions/main.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/main.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "main.Chirp": {
            "description": "A chirp created by a user",
            "type": "object",
            "properties": {
                "body": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
        "main.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "description": "Error message describing what went wrong",
                    "type": "string"
                }
            }
        },
...
        "main.requestBody": {
            "type": "object",
            "properties": {
                "body": {
                    "description": "Body is the text content of the chirp\nmax length: 140 characters",
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}